Eliminate CLS when rendering (#581)
This is a bit of a hack to push the footer down below the viewport in the first render pass. There are likely more elegant solutions, but as far as I can tell all of our content is longer than one viewport size so this won't have an effect on users. The worst case is that there's a little bit of extra whitespace and you have to scroll to see the footer.